profile update

+++ app/Http/Controllers/ProfileController.php
@@ -43,39 +43,50 @@ class ProfileController extends Controller
     	$location = $request->input('location');
 
 
-        $dir = public_path() . '\img\\';
-        $av_dir = public_path() . '\img\avatar\\';
+
+        if(!empty($_FILES['profile-pic']['name'])){
+            $dir = public_path() . '\img\\';
+            $av_dir = public_path() . '\img\avatar\\';
 
 
 
-        $target_file = $dir . basename($_FILES['profile-pic']['name']);
-        $target_file_av = $av_dir . basename($_FILES['profile-pic']['name']);
+            $target_file = $dir . basename($_FILES['profile-pic']['name']);
+            $target_file_av = $av_dir . basename($_FILES['profile-pic']['name']);
 
-  
-        if(file_exists($target_file)){  
-            return redirect()->route('profile.edit')->with('info','The file already exists!');
-        }
+            if(file_exists($target_file)){  
+                        Auth::user()->update([
+                        'name' => $request->input('name'),
+                        'location' => $request->input('location'),
+                        'profile' => '/'.$_FILES['profile-pic']['name']
+
+                    ]);
+                return redirect()->route('profile.edit')->with('info','Updated successfully!');
+            }
+            
+            move_uploaded_file($_FILES['profile-pic']['tmp_name'], $target_file);
+            copy($target_file,$target_file_av);
         
 
-
-        move_uploaded_file($_FILES['profile-pic']['tmp_name'], $target_file);
-        copy($target_file,$target_file_av);
-    
+            Image::make($target_file)->fit(80, 80)->save($target_file)->destroy();
+            Image::make($target_file_av)->fit(20, 20)->save($target_file_av)->destroy();
 
 
 
 
-        Image::make($target_file)->fit(80, 80)->save($target_file)->destroy();
-        Image::make($target_file_av)->fit(20, 20)->save($target_file_av)->destroy();
+        Auth::user()->update([
+            'name' => $request->input('name'),
+            'location' => $request->input('location'),
+            'profile' => '/'.$_FILES['profile-pic']['name']
 
+        ]);
+    }
+    else{
 
     	Auth::user()->update([
     		'name' => $request->input('name'),
-    		'location' => $request->input('location'),
-            'profile' => '/'.$_FILES['profile-pic']['name']
-
+    		'location' => $request->input('location')
     	]);
-
+    }
     	return redirect()->route('profile.edit')->with('info','Your profile has been updated!');
 
     }
